[IMP] Better modularity and naming convention for account_journal_period_close
