+++ account_journal_period_close/__openerp__.py
@@ -52,7 +52,7 @@ a given period, while leaving the miscellaneous journal open.
 From a technical standpoint, the module leverages the
 account.journal.period model that is present in Odoo core.
 """,
-    "data": ['view/account_view.xml'],
+    "data": ['view/account_period_view.xml'],
     "demo": [],
     "test": [],
     "licence": "AGPL-3",
