+++ account_journal_period_close/view/account_period_view.xml
